+++ outputs.tf
@@ -14,8 +14,16 @@ output "memorystore_instance_cpu_utilization" {
   value = { for k, v in google_monitoring_alert_policy.memorystore_instance_cpu : k => v.name }
 }
 
+output "memorystore_instance_memory_utilization" {
+  value = { for k, v in google_monitoring_alert_policy.memorystore_instance_memory : k => v.name }
+}
+
 output "memorystore_cluster_cpu_utilization" {
   value = { for k, v in google_monitoring_alert_policy.memorystore_cluster_cpu : k => v.name }
+}
+
+output "memorystore_cluster_memory_utilization" {
+  value = { for k, v in google_monitoring_alert_policy.memorystore_cluster_memory : k => v.name }
 }
 
 output "ssl_alert_policy_names" {